3376: Moved getBreadcrumbs back to breadcrumbs file
